# docs/swagger.json
{
	"components": {
...
		"schemas": {
			"AdminType": {
				"enum": [
					"super_admin",
					"admin",
					"ads_viewer",
					"store_viewer"
				],
				"type": "string"
			},
			"AdminsResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"adminType": {
						"$ref": "#/components/schemas/AdminType"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"email",
					"phone",
					"name",
					"adminType",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminPwResponse": {
				"properties": {
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminRegister": {
				"properties": {
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adminType": {
						"$ref": "#/components/schemas/AdminType"
					}
				},
				"required": [
					"email",
					"phone",
					"name",
					"password",
					"adminType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"adminType": {
						"$ref": "#/components/schemas/AdminType"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"email",
					"phone",
					"name",
					"adminType",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminPatch": {
				"properties": {
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adminType": {
						"$ref": "#/components/schemas/AdminType"
					}
				},
				"required": [
					"email",
					"phone",
					"name",
					"password",
					"adminType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminPatchPw": {
				"properties": {
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserType": {
				"enum": [
					"admin",
					"user",
					"branch",
					"owner",
					"bo"
				],
				"type": "string"
			},
			"UserStatus": {
				"enum": [
					"active",
					"quit",
					"hold"
				],
				"type": "string"
			},
			"Gender": {
				"enum": [
					"male",
					"female"
				],
				"type": "string"
			},
...
			"AuthAdminResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adminType": {
						"$ref": "#/components/schemas/AdminType"
					},
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email",
					"phone",
					"name",
					"password",
					"adminType",
					"accessToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthAdminRegister": {
				"properties": {
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adminType": {
						"$ref": "#/components/schemas/AdminType"
					}
				},
				"required": [
					"email",
					"phone",
					"name",
					"password",
					"adminType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userid": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"phoneVerified": {
						"type": "boolean"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/UserType"
					},
					"status": {
						"$ref": "#/components/schemas/UserStatus"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"avatar": {
						"type": "string"
					},
					"accessToken": {
						"type": "string"
					},
					"pushToken": {
						"type": "string"
					},
					"appleToken": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userid",
					"email",
					"phone",
					"phoneVerified",
					"emailVerified",
					"name",
					"password",
					"userType",
					"status",
					"gender",
					"avatar",
					"accessToken",
					"pushToken",
					"appleToken",
					"createdAt",
					"updatedAt",
					"deletedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access_token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "api-server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/admin/userlist": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AdminsResponse"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/admin/password/{id}": {
			"get": {
				"operationId": "ReadPassword",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminPwResponse"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdatePw",
				"responses": {
					"204": {
						"description": ""
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminPatchPw"
							}
						}
					}
				}
			}
		},
		"/admin/admin/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"204": {
						"description": ""
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminRegister"
							}
						}
					}
				}
			}
		},
		"/admin/admin/{id}": {
			"get": {
				"operationId": "Read",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminResponse"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": ""
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminPatch"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": ""
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/auth/admin/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthAdminResponse"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthAdminRegister"
							}
						}
					}
				}
			}
		},
...
		"/auth/user/userlist": {
			"get": {
				"operationId": "Userlist",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/User"
											},
											"type": "array"
										}
									},
									"required": [
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}